Remove splashscreen generation from IImageEncoder and add IImageGenerator
