+++ MediaBrowser.Controller/Drawing/GeneratedImages.cs
@@ -1,0 +1,13 @@
+namespace MediaBrowser.Controller.Drawing
+{
+    /// <summary>
+    /// Which generated images an <see cref="IImageGenerator"/> supports.
+    /// </summary>
+    public enum GeneratedImages
+    {
+        /// <summary>
+        /// The splashscreen.
+        /// </summary>
+        Splashscreen
+    }
+}
